Move the list of tags to Markapl::Tags package.
Also define a list of HTML tags that are named the same as any of the
perl builtins or ops. In those cases, an "html_" prefix is required to
output those tags. For example: html_q output <q>, html_sub outpus
<sub>. This should be better then actually clobbering them, then having
to use CORE:: if it's really wanted.

Well, while in the view code, one sane programmer shouldn't really
call "link" to deal with files, so maybe I'll change my mind sometime.