Rewrite the attribute passing part so it's not require PadWalker and other modules. Much more sane (in D::D 's sense, I guess.)
So when it sees:

   a(href => $url) { "..." }

Now it emits

   a href => $url, sub { " ... }

And now the shadow sub doesn't have prototype so this can works.

One caveats is when you pass template args as hash, hash migh clash
with tag names and cause parsing error.

template foo => sub {
    my ($self, %var) = @_;
    h1 { $var{title} }
}

Because, that emits:

template foo => sub {
    my ($self, %var) = @_;
    h1 sub { $varsub{title} } # cause syntax error
}

for some reason. not sure how to fix this yet.