Solve the problem where
   $args{title}

emits

   $argssub{title}

in the shadow sub.

In the case where tag name are used as hash key, at the point it
calls _handle_tag, the processed line offset can be indicated
like this (let pipe char mean a zero-width cursor):

   $args|{title}

And then, after skip_declarator, the $Offset still remain the same.

While if it's used to generate tags, it goes from here:

  |h1 { "..." }

to here:

  h1| { "..." }

This property is then used to tell if that declarator is used in the
context of generating tags. Need to find other extreme cases to prove
it.